backend/requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.1
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
Flask-Migrate>=4.0.7
Flask-JWT-Extended>=4.6.0
Flask-Cors>=4.0.1
